refactor(pipeline): extract build_symbol_table as a free function
