+++ codeanalyzer/core.py
@@ -8,7 +8,6 @@ from typing import Any, Dict, Optional, Union, List
 
 import time
 
-import ray
 from codeanalyzer.utils import logger
 from codeanalyzer.schema import (
     Analysis,
@@ -36,46 +35,6 @@ from codeanalyzer.syntactic_analysis.symbol_table_builder import SymbolTableBuil
 from codeanalyzer.utils import ProgressBar
 from codeanalyzer.options import AnalysisOptions
 from codeanalyzer.provenance import analyzer_info, repository_info
-
-def _ensure_ray() -> None:
-    """Initialize Ray with the driver's pinned hash seed in the workers.
-
-    An implicit auto-init would not carry PYTHONHASHSEED into worker
-    interpreters, so PyCG shards (and Jedi inference) run there with random
-    set-iteration order and the emitted edges vary run to run (issue #99)."""
-    if not ray.is_initialized():
-        ray.init(
-            runtime_env={
-                "env_vars": {
-                    "PYTHONHASHSEED": os.environ.get("PYTHONHASHSEED", "0")
-                }
-            },
-        )
-
-
-@ray.remote
-def _process_file_with_ray(py_file: Union[Path, str], project_dir: Union[Path, str], virtualenv: Union[Path, str, None]) -> Dict[str, PyModule]:
-    """Processes files in the project directory using Ray for distributed processing.
-    
-    Args:
-        py_file (Union[Path, str]): Path to the Python file to process.
-        project_dir (Union[Path, str]): Path to the project directory.
-        virtualenv (Union[Path, str, None]): Path to the virtual environment directory.
-    Returns:
-        Dict[str, PyModule]: A dictionary mapping file paths to PyModule objects.
-    """
-    from rich.console import Console
-    console = Console()
-    module_map: Dict[str, PyModule] = {}
-    try:
-        py_file = Path(py_file)
-        symbol_table_builder = SymbolTableBuilder(project_dir, virtualenv)
-        module_map[str(py_file.relative_to(Path(project_dir)))] = symbol_table_builder.build_pymodule_from_file(py_file)
-    except Exception as e:
-        console.log(f"❌ Failed to process {py_file}: {e}")
-        raise SymbolTableBuilderRayError(f"Ray processing error for {py_file}: {e}")
-    return module_map
-
 
 class Codeanalyzer:
     """Core static analysis engine for Python projects.
@@ -755,35 +714,6 @@ class Codeanalyzer:
 
         logger.info(f"Analysis cached to {cache_file}")
 
-    def _file_unchanged(self, file_path: Path, cached_module: PyModule) -> bool:
-        """Check if a file has changed since it was cached.
-        
-        Args:
-            file_path: Path to the file to check
-            cached_module: The cached PyModule for this file
-            
-        Returns:
-            bool: True if file is unchanged, False otherwise
-        """
-        try:
-            # Check last modified time and file size
-            if (cached_module.last_modified is not None and
-                cached_module.file_size is not None and
-                cached_module.last_modified == file_path.stat().st_mtime and
-                cached_module.file_size == file_path.stat().st_size):
-                return True
-            # Also check content hash for extra safety
-            if cached_module.content_hash is not None:
-                content_hash = hashlib.sha256(file_path.read_bytes()).hexdigest()
-                return content_hash == cached_module.content_hash
-
-            # No cached metadata mismatch, assume file changed
-            return False
-            
-        except Exception as e:
-            logger.debug(f"Error checking file {file_path}: {e}")
-            return False
-
     def _compute_checksum(self, root: Path) -> str:
         """Compute SHA256 checksum of all Python source files in a project directory. If somethings changes, the
         checksum will change and thus the analysis will be redone.
@@ -799,143 +729,11 @@ class Codeanalyzer:
             sha256.update(py_file.read_bytes())
         return sha256.hexdigest()
 
-    def _build_symbol_table(self, cached_symbol_table: Optional[Dict[str, PyModule]] = None) -> Dict[str, PyModule]:
-        """Builds the symbol table for the project.
-
-        This method scans the project directory, identifies Python files,
-        and constructs a symbol table containing information about classes,
-        functions, and variables defined in those files.
-        
-        Args:
-            cached_app: Previously cached PyApplication to reuse unchanged files
-        
-        Returns:
-            Dict[str, PyModule]: A dictionary mapping file paths to PyModule objects.
-        """
-        symbol_table: Dict[str, PyModule] = {}
-        t0_st = time.perf_counter()
-
-        # Handle single file analysis
-        if self.file_name is not None:
-            single_file = self.project_dir / self.file_name
-            logger.info(f"Analyzing single file: {single_file}")
-
-            # Check if file is in cache and unchanged
-            file_key = str(single_file.relative_to(self.project_dir))
-            if file_key in cached_symbol_table and not self.rebuild_analysis:
-                # Compute file checksum to see if it changed
-                if self._file_unchanged(single_file, cached_symbol_table[file_key]):
-                    logger.info(f"Using cached analysis for {single_file}")
-                    symbol_table[file_key] = cached_symbol_table[file_key]
-                    return symbol_table
-            
-            # File is new or changed, analyze it
-            try:
-                symbol_table_builder = SymbolTableBuilder(self.project_dir, self.virtualenv)
-                py_module = symbol_table_builder.build_pymodule_from_file(single_file)
-                symbol_table[file_key] = py_module
-                logger.info("✅ Single file analysis complete.")
-                return symbol_table
-            except Exception as e:
-                logger.error(f"Failed to process {single_file}: {e}")
-                return symbol_table
-        
-        # Get all Python files first to show accurate progress
-        py_files = []
-        for py_file in self.project_dir.rglob("*.py"):
-            rel_path = py_file.relative_to(self.project_dir)
-            path_parts = rel_path.parts
-            filename = py_file.name
-
-            # Skip directories we don't care about
-            if (
-                "site-packages" in path_parts
-                or ".venv" in path_parts
-                or ".codeanalyzer" in path_parts
-            ):
-                continue
-
-            # Skip test files if enabled
-            if self.skip_tests and (
-                "test" in path_parts
-                or "tests" in path_parts
-                or filename.startswith("test_")
-                or filename.endswith("_test.py")
-            ):
-                continue
-
-            py_files.append(py_file)
-
-        if self.using_ray:
-            logger.info("Using Ray for distributed symbol table generation.")
-            # Separate files into cached and new/changed
-            files_to_process = []
-            for py_file in py_files:
-                file_key = str(py_file.relative_to(self.project_dir))
-                if file_key in cached_symbol_table and not self.rebuild_analysis:
-                    if self._file_unchanged(py_file, cached_symbol_table[file_key]):
-                        # Use cached version
-                        symbol_table[file_key] = cached_symbol_table[file_key]
-                        continue
-                files_to_process.append(py_file)
-            
-            # Process only new/changed files with Ray
-            if files_to_process:
-                _ensure_ray()
-                futures = [_process_file_with_ray.remote(py_file, self.project_dir, str(self.virtualenv) if self.virtualenv else None) for py_file in files_to_process]
-                
-                with ProgressBar(len(futures), "Building symbol table (parallel)") as progress:
-                    pending = futures[:]
-                    while pending:
-                        done, pending = ray.wait(pending, num_returns=1)
-                        result = ray.get(done[0])
-                        if result:
-                            symbol_table.update(result)
-                        progress.advance()
-        else:
-            logger.info("Building symbol table serially.")
-            symbol_table_builder = SymbolTableBuilder(self.project_dir, self.virtualenv)
-            files_processed = 0
-            files_from_cache = 0
-            
-            with ProgressBar(len(py_files), "Building symbol table") as progress:
-                for py_file in py_files:
-                    file_key = str(py_file.relative_to(self.project_dir))
-                    
-                    # Check if file is cached and unchanged
-                    if file_key in cached_symbol_table and not self.rebuild_analysis:
-                        if self._file_unchanged(py_file, cached_symbol_table[file_key]):
-                            symbol_table[file_key] = cached_symbol_table[file_key]
-                            files_from_cache += 1
-                            progress.advance()
-                            continue
-                    
-                    # File is new or changed, analyze it
-                    try:
-                        py_module = symbol_table_builder.build_pymodule_from_file(py_file)
-                        symbol_table[file_key] = py_module
-                        files_processed += 1
-                    except Exception as e:
-                        logger.error(f"Failed to process {py_file}: {e}")
-                    progress.advance()
-            
-            if files_from_cache > 0:
-                logger.info(f"Reused {files_from_cache} files from cache, processed {files_processed} new/changed files")
-
-        if py_files and not symbol_table:
-            logger.error(
-                "Every one of the %d discovered Python files failed to process — "
-                "the symbol table is empty. This usually means the analysis "
-                "environment's interpreter is newer than the installed jedi/parso "
-                "stack supports (#107); check the per-file errors above.",
-                len(py_files),
-            )
-
-        logger.info(
-            "✅ Symbol table: %d modules in %.1fs",
-            len(symbol_table), time.perf_counter() - t0_st,
+    def _build_symbol_table(self, cached_symbol_table=None):
+        from codeanalyzer.pipeline.symbol_table import build_symbol_table
+        return build_symbol_table(
+            self.project_dir, self.virtualenv, self.options, cached_symbol_table or {}
         )
-        return symbol_table
 
     def _get_pycg_call_graph(
         self,
